requirements.txt
aiogram
python-whois
python-dotenv
aiosqlite